:-1:
And lo, led by the valiant hero Angus McFife XIII
The forces of justice assembled their armies in the skies above Mars
In preparation for the epic battle against the demon horde
But on planet Earth, a far more sinister machination was afoot
In the dwarven caverns beneath the mighty citadel of Dundee
The evil wizard Zargothrax began to recite the dread incantation
Which would unlock the Chaos Portal to the galactic nexus
As foretold in the dark prophecy of Anstruther countless centuries ago
As he placed the Goblin King's crystal key into the altar before him
Ancient runes began to glow on the surface of the portal
Soon the gateway would open and the elder god Kor-Virliath
Of the eighteenth hell dimension would be unleashed onto the galaxy
The countdown to universal annihilation had begun!
Warriors of planet Earth, hear my raging cry
For mighty Dundee, the demons will die
Yes, battlecruisers
To space we'll go
Discover new worlds
And conquer galaxies
Across the skies of Mars, the demon war begins
Trust in your sword, this battle we'll win
The Space Knights of Crail are first to the fight
Clashing with lasers and power of might
Fly high through apocalypse skies
Fight for the world we must save
Like tears of a unicorn lost in the rain
Chaos will triumph this day
Apocalypse 1992
At the speed of light the dwarven king arrives
With crystal laser battleaxe into the fight he rides
Pathetic dwarves!
You are no match for my demon hordes
Let battle commence!
It's the rage, the cosmic rage
The cosmic rage of astral dwarves from Aberdeen
From their mines they will arise and fight
The rage of the dwarves is tonight
It's the rage, the cosmic rage
The cosmic rage of astral dwarves from Aberdeen
From their mines they will arise and fight
The rage of the dwarves is tonight
De mortuis crepitus
Aberdonensis de furor
Infernus irae!
Then they died (mwah-ha-ha-ha!)
My ancestral demon army
Will ride a cosmic sphere
And liberate the multiverse
From slavery and fear
With the power of the crystal
From an ancient galaxy
The force of evil will prevail
It is my destiny
Fly high through apocalypse skies
Fight for the world we must save
Like tears of a unicorn lost in the rain
Chaos will triumph this day
Fly high through apocalypse skies
Fight for the world we must save
Like tears of a unicorn lost in the rain
Chaos will triumph this day
Apocalypse 1992
From the caves beneath Dundee
Ancient hermit arrives
A messenger to the war in the stars
Kor-Virliath is nigh!
Deep inside the Hootsman there lies a secret heart
Barbarian is a cyborg powered by a neutron star
For centuries immortal he will quest in time and space
But now he must make sacrifice to save the human race!
The only way to save the galaxy
Is to destroy planet Earth and Aberdeen
Faster than a laser bullet
Hootsman flies to Fife
He detonates his nuclear heart
Destroys all human life
Fly high through apocalypse skies
Fight for the world we must save
Like tears of a unicorn lost in the rain
Chaos will triumph this day
Fly high through apocalypse skies
Fight for the world we must save
Like tears of a unicorn lost in the rain
Chaos will triumph this day
Apocalypse
With a thunderous implosion the Earth was vaporised
Tearing a dimensional rift in the heavens
Using his last shards of power
Zargothrax plunged through the space portal
Vanishing into another reality
And so Angus McFife XIII followed him into the wormhole
For the eternal glory of Dundee!
Sanctus Dominus